fix: fixes and cleanup after blueprint separation
